Move navbar logic from App component to it's own component
