+++ src/Components/Navigation/Navbar/NavDesktop.js
@@ -1,10 +1,10 @@
 import React from "react";
-import "../../Styles/Navigation/Navbar.css"
 import NavBrand from "./NavBrand.js";
 import NavMenu from "./NavMenu.js";
+import "../../../Styles/Navigation/Navbar/Desktop.css";
 
 
-export default function Navbar(){
+export default function NavDesktop(){
     let nav_menu_items = {Categories: "/", Write: "/write", "Sign in": "/help", "Sign Up": "/about"};
 
     return(
