+++ src/Components/Navigation/Navbar/NavMenu.js
@@ -1,0 +1,10 @@
+import NavList from "../../Lists/NavList";
+
+export default function NavMenu(props){
+
+    return(
+        <nav id="nav-menu">
+            <NavList items={ props.items } direction={ props.direction } onItemClick={ props.onItemClick }/>
+        </nav>
+    );
+}
